Integrate Arkham to mainline
BZ: 127460

Arkham is a deeply integrated security container capability for
Android for the purpose of isolation and management in mixed used
cases. Primarily Arkham is intended to enable separation of work and
personal content and capability.

Detailed information about the Arkham requirements and architecture can
be found here: https://sharepoint.amr.ith.intel.com/sites/SDI/Arkham/

This patch modifies the KeyChainService so that containers can store
the KeyChain information separately from the primary user. To actual
work is done in vendor/intel/arkahm in the ParentyKey, this is just a
hook patch.

Change-Id: I336877eb48c23ac18dfbf3256eda3eaaf6d206c9
Signed-off-by: Octavian Purdila <octavian.purdila@intel.com>
Reviewed-on: http://android.intel.com:8080/126924
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Shuster, Norvin <norvin.shuster@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>